+++ app/server/modules/email-dispatcher.js
@@ -17,11 +17,11 @@ const composeActivationEmailViewDefault = require(
 
 /**
 * @typedef {PlainObject} EmailDispatcherConfig
-* @property {string} [NL_EMAIL_HOST='smtp.gmail.com']
-* @property {string} NL_EMAIL_USER E.g., your-email-address@gmail.com
+* @property {string} NL_EMAIL_HOST
+* @property {string} NL_EMAIL_USER E.g., your-email-address@example.name
 * @property {string} NL_EMAIL_PASS E.g., 1234
 * @property {string} NL_EMAIL_FROM
-* @property {string} [NL_SITE_URL='http://localhost:3000']
+* @property {string} NL_SITE_URL
 */
 
 /**
@@ -34,12 +34,12 @@ class EmailDispatcher {
    */
   constructor (config) {
     const {
-      NL_EMAIL_HOST = 'smtp.gmail.com',
+      NL_EMAIL_HOST,
       NL_EMAIL_USER, // = 'your-email-address@example.name',
       NL_EMAIL_PASS, // = '1234',
       NL_EMAIL_FROM, // = 'Node Login <do-not-reply@example.name>',
       NS_EMAIL_TIMEOUT,
-      NL_SITE_URL = 'http://localhost:3000',
+      NL_SITE_URL,
       composeResetPasswordEmailView = composeResetPasswordEmailViewDefault,
       composeActivationEmailView = composeActivationEmailViewDefault
     } = config;