+++ nogin-sample.js
@@ -6,5 +6,6 @@ module.exports = {
   NL_EMAIL_USER: 'me@example.name',
   NL_EMAIL_PASS: 'abc123456',
   NL_EMAIL_FROM: 'Node Login <do-not-reply@example.name>',
+  NL_SITE_URL: 'https://example.name',
   NS_EMAIL_TIMEOUT: 50000
 };