+++ CHANGES.md
@@ -2,13 +2,16 @@
 
 ## ?
 
+- Breaking change: Avoid defaults for `NL_EMAIL_HOST` and `NL_SITE_URL`.
 - License: Add `core-js` to bundled (MIT)
 - Fix: Add Rollup with Babel (and core-js) to ensure syntax working on
     older browsers; add options `useESM` and `noPolyfill` for configuring
 - Update: Global setting no longer needed for `intl-dom`
+- Enhancement: i18nize query params (redirect, key, c)
 - Linting: Enforce `no-restricted-globals` with `window`
 - Linting: Avoid disabling compat plugin for client-side code
 - Docs: Add migration guide
+- Docs: Update CLI help
 - npm: Move `chai-as-promised` to devDeps and `nodemon`/`npm-run-all`
     to deps.
 - npm: Add `license-badges` to `prepare` script